Fix Cucumber issues from merge, rubocop

+++ db/migrate/20241113211539_drop_sections.rb
@@ -1,24 +1,17 @@
 class DropSections < ActiveRecord::Migration[7.2]
   def change
-    if foreign_key_exists?(:room_bookings, :sections)
-      remove_foreign_key :room_bookings, :sections
-    end
+    remove_foreign_key :room_bookings, :sections if foreign_key_exists?(:room_bookings, :sections)
 
-    if index_exists?(:room_bookings, :section_id)
-      remove_index :room_bookings, :section_id
-    end
+    remove_index :room_bookings, :section_id if index_exists?(:room_bookings, :section_id)
 
-    if column_exists?(:room_bookings, :section_id)
-      remove_column :room_bookings, :section_id, :integer
-    end
+    remove_column :room_bookings, :section_id, :integer if column_exists?(:room_bookings, :section_id)
 
     # Get rid of sections
-    if table_exists?(:sections)
-      drop_table :sections do |t|
-        t.references :course, foreign_key: true
-        t.timestamps
-      end
+    return unless table_exists?(:sections)
+
+    drop_table :sections do |t|
+      t.references :course, foreign_key: true
+      t.timestamps
     end
   end
-  
 end
